COM-1923 - use exitConfirmationModal in card store instead of activities store
