+++ src/components/cards/CardHeader/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { View } from 'react-native';
 import { connect } from 'react-redux';
 import FeatherButton from '../../icons/FeatherButton';
-import Actions from '../../../store/activities/actions';
+import Actions from '../../../store/cards/actions';
 import { ICON } from '../../../styles/metrics';
 import { GREY } from '../../../styles/colors';
 import { ActionType, StateType } from '../../../types/store/StoreType';
