+++ src/store/activities/reducers.ts
@@ -1,7 +1,6 @@
 import {
   ActivityStateType,
   SET_ACTIVITY,
-  SET_EXIT_CONFIRMATION_MODAL,
   RESET_ACTIVITY_REDUCER,
   SET_ACTIVITY_HISTORIES,
   ActivityActionType,
@@ -10,7 +9,6 @@ import {
 
 const initialState: ActivityStateType = {
   activity: null,
-  exitConfirmationModal: false,
   activityHistories: [],
 };
 
@@ -21,8 +19,6 @@ export const activities = (
   switch (action.type) {
     case SET_ACTIVITY:
       return { ...state, activity: action.payload };
-    case SET_EXIT_CONFIRMATION_MODAL:
-      return { ...state, exitConfirmationModal: action.payload };
     case SET_ACTIVITY_HISTORIES:
       return { ...state, activityHistories: action.payload };
     case RESET_ACTIVITY_REDUCER:
