+++ src/types/store/CardStoreType.ts
@@ -7,6 +7,7 @@ export const ADD_QUESTIONNAIRE_ANSWER = 'ADD_QUESTIONNAIRE_ANSWER';
 export const REMOVE_QUESTIONNAIRE_ANSWER = 'REMOVE_QUESTIONNAIRE_ANSWER';
 export const SET_QUESTIONNAIRE_ANSWERS_LIST = 'SET_QUESTIONNAIRE_ANSWERS_LIST';
 export const INC_GOOD_ANSWERS_COUNT = 'INC_GOOD_ANSWERS_COUNT';
+export const SET_EXIT_CONFIRMATION_MODAL = 'SET_EXIT_CONFIRMATION_MODAL';
 
 // STATE
 export interface CardStateType {
@@ -14,6 +15,7 @@ export interface CardStateType {
   cardIndex: number | null,
   questionnaireAnswersList: Array<QuestionnaireAnswerType>,
   score: number,
+  exitConfirmationModal: boolean,
 }
 
 // ACTION
@@ -53,11 +55,17 @@ export interface IncGoodAnswersCountType {
   type: typeof INC_GOOD_ANSWERS_COUNT,
 }
 
+export interface SetExitConfirmationModalType {
+  type: typeof SET_EXIT_CONFIRMATION_MODAL,
+  payload: boolean,
+}
+
 export type CardActionType =
 SetCardsType |
 SetCardIndexType |
 AddQuestionnaireAnswerType |
 SetQuestionnaireAnswersListType |
-RemoveQuestionnaireAnswerType;
+RemoveQuestionnaireAnswerType |
+SetExitConfirmationModalType;
 
 export type CardActionWithoutPayloadType = ResetCardReducerType | IncGoodAnswersCountType;
